Change routing file structure
