+++ src/app/routes/loader/Loader.module.scss
